Consolidation
Removing the logic for testing STDIN and leveraging an upstream library
seems like a good idea. Collapsing all of the functions into something
considerably more imperative, maybe less so. I worry that this makes
the behavior of the program harder to test, but the goal is understanding
how to write tests in Go so... win/win?